Change binary install script to handle all app pkgs
- Also removes extraneous files on install
- Addresses #6, #11, #15

+++ add-cran-binary-pkgs.R
@@ -1,3 +1,5 @@
+#!/usr/bin/env Rscript
+
 # Script to find dependencies of a pkg list, download binaries and put them
 # In the standalone R library.
 
@@ -8,7 +10,9 @@ cran_pkgs <- unique(c(
   automagic::get_dependent_packages("shiny")
 ))
 
-install_bins <- function(cran_pkgs, library_path, type, decompress) {
+install_bins <- function(cran_pkgs, library_path, type, decompress,
+                         remove_dirs = c("help", "doc", "tests", "html", "include")) {
+  
   installed <- list.files(library_path)
   cran_to_install <- sort(setdiff(
     unique(unlist(
@@ -16,21 +20,28 @@ install_bins <- function(cran_pkgs, library_path, type, decompress) {
         tools::package_dependencies(cran_pkgs, recursive=TRUE,
                                     which= c("Depends", "Imports", "LinkingTo"))))),
     installed))
-  td <- tempdir()
-  downloaded <- download.packages(cran_to_install, destdir = td, type=type)
-  apply(downloaded, 1, function(x) decompress(x[2], exdir = library_path))
-  unlink(downloaded[,2])
+  if(!length(cran_to_install)) {
+    message("No packages to install")
+    return(invisible(NULL))
+  } else {
+    td <- tempdir()
+    downloaded <- download.packages(cran_to_install, destdir = td, type=type)
+    apply(downloaded, 1, function(x) decompress(x[2], exdir = library_path))
+    unlink(downloaded[,2])
+  }
+  z <- lapply(list.dirs(library_path, full.names = TRUE, recursive = FALSE), 
+              function(x) {
+                unlink(file.path(x, remove_dirs), force=TRUE, recursive=TRUE)
+              })
+  invisible(NULL)
 }
 
 if (dir.exists("r-mac")) {
-  install_bins(cran_pkgs = "shiny", library_path = file.path("r-mac", "library"),
+  install_bins(cran_pkgs = cran_pkgs, library_path = file.path("r-mac", "library"),
                type = "mac.binary.el-capitan", decompress = untar)
 }
 
 if (dir.exists("r-win")) {
-  install_bins(cran_pkgs = "shiny", library_path = file.path("r-win", "library"),
+  install_bins(cran_pkgs = cran_pkgs, library_path = file.path("r-win", "library"),
                type = "win.binary", decompress = unzip)
 }
-
-#TODO see what parts of R package files we can remove for space,
-#Possibly help/ doc/ tests/ html/ subdirs
